Initial check-in of the Layered Windows plugin.
The layers plugin mimics tabbed browsing. It logically groups windows together and stacks them on top of each other (layer-by-layer) at the same position and with the same size. By keeping all but one window from each group invisible the taskbar stays tidy and you get the impression of only having few browser windows with several views.

Current features include opening and closing new "layers" in a number of ways. The current layers can be shown in a menu or on a buttonbar. Quick navigation with accelerator keys is also possible.