New feature: A right-click context menu named LayerButtonPopup.
This is an experimental/temporary solution that adds a very limited
menus.cfg parser to the layers plugin. Only layers() menu entries are
possible at this point.

The permanent solution ought to be a way to receive popup menus from
the main menus.cfg parser. That way we can use command ID's and also
call other plugins or open sub-menus.